move ckptmgr-client into stmgr directory
