+++ heron/stmgr/src/cpp/manager/ckptmgr-client.h
