+++ heron/stmgr/src/cpp/manager/stmgr-server.h
@@ -24,7 +24,7 @@
 #include "proto/messages.h"
 #include "network/network.h"
 #include "basics/basics.h"
-#include "client/ckptmgr-client.h"
+#include "manager/ckptmgr-client.h"
 
 namespace heron {
 namespace common {
